zmiana z int na byte

+++ palindrom.cpp
@@ -6,7 +6,7 @@ using namespace std;
 int main()
 {
     cout << "hello friends" << endl;
-    int* tablica;
+    byte* tablica;
     cout << &tablica << endl;
     cout << &tablica + 1 << endl;
     return 0;
